implemented suggestions.
